+++ src/Shrun/Data/FilePathDefault.hs
@@ -10,6 +10,7 @@ module Shrun.Data.FilePathDefault
   )
 where
 
+import Data.Char qualified as Ch
 import Shrun.Prelude
 
 -- | FilePath option that includes a default possibility.
@@ -32,7 +33,9 @@ makePrisms ''FilePathDefault
 
 -- | @since 0.5
 instance DecodeTOML FilePathDefault where
-  tomlDecoder =
-    tomlDecoder @Text <&> \case
-      "default" -> FPDefault
-      other -> FPManual $ unpack other
+  tomlDecoder = do
+    f <- tomlDecoder
+    case fmap Ch.toLower f of
+      "default" -> pure FPDefault
+      "" -> fail "Empty path given for --file-log"
+      _ -> pure (FPManual f)